src/backend/requirements.txt
# ============================================================================
# RiceGuard Backend Requirements (FastAPI + MongoDB + ML)
# ============================================================================

# ----------------------------------------------------------------------------
# Core web stack
# ----------------------------------------------------------------------------
fastapi==0.103.2
uvicorn[standard]==0.23.2
starlette==0.27.0

# ----------------------------------------------------------------------------
# Validation & request parsing
# ----------------------------------------------------------------------------
pydantic[email]==1.10.19
email-validator==2.1.0.post1
python-multipart==0.0.9

# ----------------------------------------------------------------------------
# Configuration / environment handling
# ----------------------------------------------------------------------------
python-dotenv==1.0.1

# ----------------------------------------------------------------------------
# Database (MongoDB Atlas compatible)
# ----------------------------------------------------------------------------
pymongo==4.15.3
dnspython>=2.6.0,<3.0    # explicit dependency for mongodb+srv URIs on new pymongo releases

# ----------------------------------------------------------------------------
# Security / authentication
# ----------------------------------------------------------------------------
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# ----------------------------------------------------------------------------
# Machine learning / image processing
# Note: TensorFlow 2.20.0 requires Python < 3.13
# For Python 3.13+, use CPU-only build or downgrade Python
# ----------------------------------------------------------------------------
tensorflow==2.20.0 ; python_version < "3.13"
tensorflow-cpu==2.20.0 ; python_version >= "3.13"
numpy==1.26.4 ; python_version < "3.13"
numpy>=2.1,<3.0 ; python_version >= "3.13"
pillow==10.4.0

# ----------------------------------------------------------------------------
# Testing / TDD infrastructure
# ----------------------------------------------------------------------------
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
mongomock==4.2.0
httpx==0.28.1
factory-boy==3.3.1
faker==30.8.1
freezegun==1.5.1
coverage==7.6.9

# ----------------------------------------------------------------------------
# Observability / utilities
# ----------------------------------------------------------------------------
requests==2.31.0
typing-extensions>=4.7.1
certifi>=2024.8.30
